requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug>=2.3.7
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow>=10.0.0
Flask-SQLAlchemy==3.1.1
python-docx==1.1.0
scikit-learn==1.3.2
nltk==3.8.1
regex==2023.10.3
Flask-Login==0.6.3

# Google Cloud & AI (optional but aligned with roadmap)
google-cloud-documentai==2.28.0
google-cloud-storage==2.16.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1

# Google Calendar API (optional)
google-api-python-client==2.137.0

# Notifications (Firebase, email)
firebase-admin==6.5.0
sendgrid==6.11.0

# Date parsing & timezone helpers
python-dateutil==2.9.0.post0
pytz==2024.1

# Security (optional hardening)
cryptography>=42.0.0

# Gemini API
google-generativeai==0.8.3
